+++ package.json
@@ -30,6 +30,8 @@
     "@eslint/js": "^9.33.0",
     "@react-router/dev": "^7.7.1",
     "@tailwindcss/vite": "^4.1.4",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/react": "^16.3.0",
     "@types/node": "^20",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
@@ -39,6 +41,7 @@
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-react": "^7.37.5",
     "globals": "^16.3.0",
+    "happy-dom": "^18.0.1",
     "husky": "^9.1.7",
     "prettier": "3.6.2",
     "tailwindcss": "^4.1.4",
